# Use private key fixtures for kubeadm unit tests