--- a/cmd/kubeadm/app/util/pkiutil/testing/BUILD
+++ b/cmd/kubeadm/app/util/pkiutil/testing/BUILD
@@ -1,0 +1,23 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["testing.go"],
+    importpath = "k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil/testing",
+    visibility = ["//visibility:public"],
+    deps = ["//cmd/kubeadm/app/util/pkiutil:go_default_library"],
+)
+
+filegroup(
+    name = "package-srcs",
+    srcs = glob(["**"]),
+    tags = ["automanaged"],
+    visibility = ["//visibility:private"],
+)
+
+filegroup(
+    name = "all-srcs",
+    srcs = [":package-srcs"],
+    tags = ["automanaged"],
+    visibility = ["//visibility:public"],
+)