--- a/cmd/kubeadm/app/util/pkiutil/testing/testdata/README.md
+++ b/cmd/kubeadm/app/util/pkiutil/testing/testdata/README.md
@@ -1,0 +1,1 @@
+Keys in this directory are generated for testing purposes only.