… cmd/kubeadm/app/cmd/BUILD
… cmd/kubeadm/app/util/pkiutil/testing/testdata/1.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/10.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/11.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/12.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/13.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/14.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/15.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/16.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/17.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/18.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/19.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/2.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/20.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/3.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/4.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/5.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/6.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/7.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/8.rsa.key
… cmd/kubeadm/app/util/pkiutil/testing/testdata/9.rsa.key