xref #98486

alternative to #98638

Private 

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

results in a ~75% speedup of kubeadm unit tests on my machine:

before:
```
time go test -race ./cmd/kubeadm/app/... -count=1
?   	k8s.io/kubernetes/cmd/kubeadm/app	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm	0.234s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/fuzzer	0.855s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/scheme	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta1	0.277s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2	0.382s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/validation	0.536s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/fuzzer	0.336s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/scheme	[no test files]
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/v1alpha1	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd	74.887s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/alpha	9.308s
?   	k8s.io/kubernetes/cmd/kubeadm/app/cmd/options	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases	0.188s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/init	56.250s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/join	0.609s [no tests to run]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/reset	0.476s
?   	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/upgrade/node	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/workflow	0.054s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/upgrade	4.927s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/util	0.581s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/componentconfigs	0.530s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/constants	0.337s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/discovery	0.404s
?   	k8s.io/kubernetes/cmd/kubeadm/app/discovery/file	[no test files]
?   	k8s.io/kubernetes/cmd/kubeadm/app/discovery/https	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/discovery/token	0.854s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/features	0.073s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/images	0.195s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/addons/dns	0.536s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/addons/proxy	1.263s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/bootstraptoken/clusterinfo	0.348s
?   	k8s.io/kubernetes/cmd/kubeadm/app/phases/bootstraptoken/node	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/certs	115.806s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/certs/renewal	23.032s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/controlplane	12.430s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/copycerts	22.606s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/etcd	0.558s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/kubeconfig	40.051s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/kubelet	0.607s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/markcontrolplane	3.931s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/patchnode	1.904s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/upgrade	61.283s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/uploadconfig	0.789s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/preflight	0.505s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util	0.850s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/apiclient	3.110s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/certs	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/config	12.294s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/config/strict	0.683s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/crypto	0.265s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/dryrun	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/etcd	1.055s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/image	0.052s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/initsystem	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/kubeconfig	0.458s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/output	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/patches	0.255s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil	28.259s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/pubkeypin	0.077s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/runtime	0.449s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/staticpod	0.661s

real	2m51.579s
user	14m32.353s
sys	1m7.601s
```

after:
```
time go test -race ./cmd/kubeadm/app/... -count=1
?   	k8s.io/kubernetes/cmd/kubeadm/app	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm	0.220s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/fuzzer	0.615s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/scheme	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta1	0.271s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/v1beta2	0.172s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm/validation	0.364s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/fuzzer	0.252s
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/scheme	[no test files]
?   	k8s.io/kubernetes/cmd/kubeadm/app/apis/output/v1alpha1	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd	11.400s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/alpha	5.724s
?   	k8s.io/kubernetes/cmd/kubeadm/app/cmd/options	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases	0.345s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/init	4.112s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/join	0.486s [no tests to run]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/reset	0.454s
?   	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/upgrade/node	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/workflow	0.074s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/upgrade	1.508s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/cmd/util	0.388s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/componentconfigs	0.585s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/constants	0.166s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/discovery	0.340s
?   	k8s.io/kubernetes/cmd/kubeadm/app/discovery/file	[no test files]
?   	k8s.io/kubernetes/cmd/kubeadm/app/discovery/https	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/discovery/token	0.986s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/features	0.082s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/images	0.272s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/addons/dns	0.689s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/addons/proxy	1.286s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/bootstraptoken/clusterinfo	0.419s
?   	k8s.io/kubernetes/cmd/kubeadm/app/phases/bootstraptoken/node	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/certs	4.831s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/certs/renewal	5.559s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/controlplane	0.842s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/copycerts	2.154s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/etcd	0.504s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/kubeconfig	1.398s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/kubelet	0.545s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/markcontrolplane	3.888s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/patchnode	1.909s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/upgrade	16.321s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/phases/uploadconfig	0.483s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/preflight	0.435s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util	0.370s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/apiclient	2.840s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/certs	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/config	14.375s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/config/strict	0.367s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/crypto	0.151s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/dryrun	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/etcd	0.712s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/image	0.058s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/initsystem	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/kubeconfig	0.272s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/output	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/patches	0.187s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil	16.039s
?   	k8s.io/kubernetes/cmd/kubeadm/app/util/pkiutil/testing	[no test files]
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/pubkeypin	0.065s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/runtime	0.116s
ok  	k8s.io/kubernetes/cmd/kubeadm/app/util/staticpod	0.301s

real	0m44.183s
user	3m10.531s
sys	0m29.574s
```

/cc @neolit123